LPS-50491 - Move liferayselenium service/helper files to portalweb/util/liferayselenium/service
