+++ portal-web/test/functional/com/liferay/portalweb/util/liferayselenium/service/SeleniumException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portal.util.liferayselenium;
+package com.liferay.portalweb.util.liferayselenium.service;
 
 /**
  * @author Michael Hashimoto
